refactor: move sectionhero cluster into @repo/shared (#1561)
