+++ app/components/Docs/components/SectionHero/sectionHero.styles.ts
@@ -1,14 +1,14 @@
-import { SectionHero } from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero";
-import {
-  Head,
-  SectionLayout,
-} from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero.styles";
 import {
   CONTENT_TYPE,
   FrontmatterProps,
 } from "@/docs/common/frontmatter/types";
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
+import { SectionHero } from "@repo/shared/components/layout/SectionHero/sectionHero";
+import {
+  Head,
+  SectionLayout,
+} from "@repo/shared/components/layout/SectionHero/sectionHero.styles";
 
 export const PADDING_Y_BOTTOM = 90;
