+++ app/components/Home/components/Section/components/SectionHero/sectionHero.styles.ts
@@ -1,4 +1,3 @@
-import { Section } from "@/components/common/Section/section";
 import { FONT } from "@databiosphere/findable-ui/lib/styles/common/constants/font";
 import { PALETTE } from "@databiosphere/findable-ui/lib/styles/common/constants/palette";
 import { bpUpMd } from "@databiosphere/findable-ui/lib/styles/common/mixins/breakpoints";
@@ -8,6 +7,7 @@ import {
   sectionGrid,
   sectionLayout,
 } from "@repo/shared/components/layout/Section/section.styles";
+import { Section } from "@repo/shared/components/layout/SectionHero/components/Section/section";
 
 export const StyledSection = styled(Section)`
   ${section};
