+++ packages/shared/components/layout/SectionHero/components/Hero/common/constants.ts
