+++ packages/shared/components/layout/SectionHero/components/Hero/common/types.ts
