+++ packages/shared/components/layout/SectionHero/components/Hero/common/utils.ts
