+++ packages/shared/components/layout/SectionHero/components/Hero/components/Defs/SmokeCircle/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+  gridSize: number;
+}
