+++ packages/shared/components/layout/SectionHero/components/Hero/components/Defs/SmokeRect/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+  gridSize: number;
+}
