+++ packages/shared/components/layout/SectionHero/components/Hero/hero.styles.ts
