+++ packages/shared/components/layout/SectionHero/components/Hero/types.ts
@@ -1,0 +1,5 @@
+export interface Props {
+  gridSize?: number;
+  height?: number;
+  width?: number;
+}
