+++ packages/shared/components/layout/SectionHero/components/Section/section.styles.ts
