+++ packages/shared/components/layout/SectionHero/components/Section/types.ts
@@ -1,0 +1,6 @@
+import { JSX } from "react";
+
+export interface Props {
+  children: (height?: number, width?: number) => JSX.Element;
+  className?: string;
+}
