+++ packages/shared/components/layout/SectionHero/types.ts
@@ -1,0 +1,9 @@
+import type { Breadcrumb } from "@databiosphere/findable-ui/lib/components/common/Breadcrumbs/breadcrumbs";
+import type { BaseComponentProps } from "@databiosphere/findable-ui/lib/components/types";
+import type { ReactNode } from "react";
+
+export interface Props extends BaseComponentProps {
+  breadcrumbs: Breadcrumb[];
+  head: ReactNode;
+  subHead: ReactNode;
+}
