+++ sites/brc-analytics/views/CalendarView/calendarView.tsx
@@ -1,9 +1,9 @@
-import { SectionHero } from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero";
 import {
   Section,
   SectionContent,
   SectionLayout,
 } from "@/components/content/content.styles";
+import { SectionHero } from "@repo/shared/components/layout/SectionHero/sectionHero";
 import { Fragment, JSX } from "react";
 import { BREADCRUMBS } from "./constants";
