+++ sites/brc-analytics/views/RoadmapView/roadmapView.tsx
@@ -1,6 +1,6 @@
-import { SectionHero } from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero";
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import { Typography } from "@mui/material";
+import { SectionHero } from "@repo/shared/components/layout/SectionHero/sectionHero";
 import { Fragment, JSX } from "react";
 import { BREADCRUMBS } from "./constants";
 import SectionRoadmap from "./content/sectionRoadmap.mdx";
